# Optimize conversion of GAP strings to Julia