--- a/CHANGES.md
+++ b/CHANGES.md
@@ -39,6 +39,7 @@
   - simpcomp
   - singular
   - zeromqinterface
+- Optimize conversion of GAP strings to Julia strings, symbols or `Vector{UInt8}`
 
 ## Version 0.12.3 (released 2025-01-01)